Ensure correct behaviour of ExecutionContextObservables
As it uses an ExecutionContext there can be a race condition between
calls of the methods.  onSubscribe and onNext need to block to ensure
that they have been processed before allowing further calls to the
Observable. onError and onComplete do not need to block as they are
by contract terminal methods on the observer.

SCALA-405